# server/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  model.AdminMessageOperateReq:
    properties:
      id:
        description: 消息主键id
        type: integer
    required:
    - id
    type: object
  model.AdminTransferGroupOwnerReq:
    properties:
      targetUserId:
        description: 目标新群主 userId
        type: string
    required:
    - targetUserId
    type: object
...
  model.UpdateChatGroup:
    properties:
      avatar:
        description: 群头像（可选）
        type: string
      groupName:
        description: 群名称（可选）
        type: string
      intro:
        description: 群介绍（可选）
        type: string
      notice:
        description: 群公告（可选）
        type: string
    type: object
...
  response.ResType:
    properties:
      code:
        default: 0
        type: integer
      data: {}
      encrypt:
        type: string
      msg:
        default: 操作成功
        type: string
    type: object
host: localhost:9600
info:
  contact: {}
  description: zone服务端API服务
  title: Gin-Zone-Api
  version: "1.0"
paths:
...
  /admin/chat/friends:
    get:
      consumes:
      - application/json
      description: 管理端-好友/群聊关系列表
      parameters:
      - in: query
        name: friendId
        type: string
      - in: query
        name: groupId
        type: integer
      - example: 1
        in: query
        name: page
        type: integer
      - example: 20
        in: query
        name: pageSize
        type: integer
      - in: query
        name: userId
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ResType'
      security:
      - Authorization: []
      summary: 管理端-好友/群聊关系列表
      tags:
      - 管理端-聊天模块
  /admin/chat/friends/{id}:
    delete:
      consumes:
      - application/json
      description: 管理端-按关系主键删除（z_chat_friends.id）
      parameters:
      - description: 关系id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ResType'
      security:
      - Authorization: []
      summary: 管理端-删除好友/群聊关系
      tags:
      - 管理端-聊天模块
  /admin/chat/groupMembers/{groupId}/{userId}:
    delete:
      consumes:
      - application/json
      description: 管理端-删除群成员（管理员）
      parameters:
      - description: 群组id
        in: path
        name: groupId
        required: true
        type: integer
      - description: 成员用户id
        in: path
        name: userId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ResType'
      security:
      - Authorization: []
      summary: 管理端-删除群成员
      tags:
      - 管理端-聊天模块
  /admin/chat/groups:
    get:
      consumes:
      - application/json
      description: 管理端-群组列表（支持 groupName 模糊查询）
      parameters:
      - in: query
        name: groupName
        type: string
      - example: 1
        in: query
        name: page
        type: integer
      - example: 20
        in: query
        name: pageSize
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ResType'
      security:
      - Authorization: []
      summary: 管理端-群组列表
      tags:
      - 管理端-聊天模块
  /admin/chat/groups/{groupId}:
    patch:
      consumes:
      - application/json
      description: 管理端-修改群聊信息（管理员）
      parameters:
      - description: 群组id
        in: path
        name: groupId
        required: true
        type: integer
      - description: 需要上传的json
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/model.UpdateChatGroup'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ResType'
      security:
      - Authorization: []
      summary: 管理端-修改群聊信息
      tags:
      - 管理端-聊天模块
  /admin/chat/groups/{id}:
    delete:
      consumes:
      - application/json
      description: 管理端-解散群组（群主或管理员）
      parameters:
      - description: 群组id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ResType'
      security:
      - Authorization: []
      summary: 管理端-解散群组
      tags:
      - 管理端-聊天模块
    get:
      consumes:
      - application/json
      description: 管理端-查询群组详情（含群主信息）
      parameters:
      - description: 群组id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ResType'
      security:
      - Authorization: []
      summary: 管理端-群组详情
      tags:
      - 管理端-聊天模块
  /admin/chat/groups/{id}/transferOwner:
    post:
      consumes:
      - application/json
      description: 管理端-转移群主（群主或管理员）
      parameters:
      - description: 群组id
        in: path
        name: id
        required: true
        type: integer
      - description: 转移参数
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/model.AdminTransferGroupOwnerReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ResType'
      security:
      - Authorization: []
      summary: 管理端-转移群主
      tags:
      - 管理端-聊天模块
  /admin/chat/messages:
    get:
      consumes:
      - application/json
      description: 管理端-按发送人/群组/关键词/时间范围/类型检索消息
      parameters:
      - description: 结束时间（yyyy-MM-dd HH:mm:ss）
        in: query
        name: endAt
        type: string
      - in: query
        name: groupId
        type: integer
      - in: query
        name: keyword
        type: string
      - in: query
        name: msgType
        type: integer
      - example: 1
        in: query
        name: page
        type: integer
      - example: 20
        in: query
        name: pageSize
        type: integer
      - in: query
        name: senderId
        type: string
      - description: 起始时间（yyyy-MM-dd HH:mm:ss）
        in: query
        name: startAt
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ResType'
      security:
      - Authorization: []
      summary: 管理端-消息检索
      tags:
      - 管理端-聊天模块
  /admin/chat/messages/delete:
    post:
      consumes:
      - application/json
      description: 管理端-软删除消息
      parameters:
      - description: 处置参数
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/model.AdminMessageOperateReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ResType'
      security:
      - Authorization: []
      summary: 管理端-删除消息
      tags:
      - 管理端-聊天模块
  /admin/chat/messages/revoke:
    post:
      consumes:
      - application/json
      description: 管理端-软撤回消息
      parameters:
      - description: 处置参数
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/model.AdminMessageOperateReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ResType'
      security:
      - Authorization: []
      summary: 管理端-撤回消息
      tags:
      - 管理端-聊天模块
...
securityDefinitions:
  Authorization:
    description: jwt token 鉴权
    in: header
    name: Authorization
    type: apiKey
  BasicAuth:
    type: basic
swagger: "2.0"